fix: : validation for location coordinates

+++ repository/src/main/java/org/example/store/repository/GeoLocationRepository.java
@@ -5,6 +5,7 @@ import org.example.store.model.GeoLocation;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import java.math.BigDecimal;
 import java.util.Optional;
 
 @Repository
@@ -14,6 +15,6 @@ public interface GeoLocationRepository extends JpaRepository<GeoLocation, Long> 
 
     Optional<GeoLocation> findByLongitude(Double longitude);
 
-    Optional<GeoLocation> findByLatitudeAndLongitude(Double latitude, Double longitude);
+    Optional<GeoLocation> findByLatitudeAndLongitude(BigDecimal latitude, BigDecimal longitude);
 
 }
